requirements.txt
discord.py
requests
schedule
 